Finally receive data from punkapi.com
The main reason for the initial problems was the weird "passing a
default value in form of an object already" to useDataApi. This would
have been better designed if it was declared inside useDataApi.
Furthermore, if we are only interested in the actual data, which is
hidden inside an object named data.hits, then why not directly access
and return that?